Rename licensing module to license_notice
I forgot that TinyPilot Pro already has a file called licensing.py that serves a different purpose.
