+++ app/main.py
@@ -12,7 +12,7 @@ from werkzeug import exceptions
 import log
 import api
 import json_response
-import licensing
+import license_notice
 import secret_key
 import socket_api
 import views
@@ -47,7 +47,7 @@ app.config.from_envvar('APP_SETTINGS_FILE')
 csrf = flask_wtf.csrf.CSRFProtect(app)
 
 app.register_blueprint(api.api_blueprint)
-app.register_blueprint(licensing.blueprint)
+app.register_blueprint(license_notice.blueprint)
 app.register_blueprint(views.views_blueprint)
