+++ requirements.txt
@@ -1,4 +1,4 @@
-# When modifying dependencies, update the credit in app/licensing.py.
+# When modifying dependencies, update the credit in app/license_notice.py.
 
 eventlet==0.31.0
 Flask==1.1.1
